[simpleSpmv] change the index type from int to size_t; add the COO matrix storage format in the CUDA, HIP and SYCL programs

+++ src/simpleSpmv-omp/kernels.cpp
@@ -6,7 +6,7 @@
 // dense matrix vector multiply
 long mv_dense_parallel(const int repeat,
                        const int bs,
-                       const int num_rows,
+                       const size_t num_rows,
                        const REAL* x,
                              REAL* matrix,
                              REAL* y)
@@ -21,9 +21,9 @@ long mv_dense_parallel(const int repeat,
 
     for (int n = 0; n < repeat; n++) {
       #pragma omp target teams distribute parallel for num_threads(bs)
-      for (int i = 0; i < num_rows; i++) {
+      for (size_t i = 0; i < num_rows; i++) {
         REAL temp = 0;
-        for (int j = 0; j < num_rows; j++) {
+        for (size_t j = 0; j < num_rows; j++) {
           if (matrix[i * num_rows + j] != (REAL)0) 
             temp += matrix[i * num_rows + j] * x[j];
         }
@@ -41,14 +41,14 @@ long mv_dense_parallel(const int repeat,
 // sparse matrix vector multiply using the CSR format
 long mv_csr_parallel(const int repeat,
                      const int bs,
-                     const int num_rows,
+                     const size_t num_rows,
                      const REAL* x,
                      const size_t nnz,
                      REAL* matrix,
                      REAL* y)
 {
   size_t *row_indices = (size_t *) malloc((num_rows+1) * sizeof(size_t));
-  int *col_indices = (int *) malloc(nnz * sizeof(int));
+  size_t *col_indices = (size_t *) malloc(nnz * sizeof(size_t));
   REAL *values = (REAL *) malloc(nnz * sizeof(REAL));
 
   // initialize csr structure
@@ -66,7 +66,7 @@ long mv_csr_parallel(const int repeat,
 
     for (int n = 0; n < repeat; n++) {
       #pragma omp target teams distribute parallel for num_threads(bs)
-      for (int i = 0; i < num_rows; i++) {
+      for (size_t i = 0; i < num_rows; i++) {
         size_t row_start = row_indices[i];
         size_t row_end = row_indices[i+1];
